requirements.txt
# requirements.txt:

pytest
selenium
requests
python-dotenv
allure-pytest
python-arango
pytest-xdist
docker
psycopg2-binary